# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberQuest AI defensive training API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  code: string;
  message: string;
}

export interface ErrorResponse {
  success: false;
  error: ApiError;
}

export type SuccessResponseData = { [key: string]: unknown };

export interface SuccessResponse {
  success: true;
  data: SuccessResponseData;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  xp: number;
  level: number;
  streak: number;
  createdAt: string;
}

export interface UserData {
  user: User;
}

export interface UserResponse {
  success: true;
  data: UserData;
}

export interface AuthResponse {
  success: true;
  data: UserData;
}

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 10
     * @maxLength 100
     */
  password: string;
}

export interface LoginInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  password: string;
}

export interface ProfileUpdateInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
}

export type Difficulty = typeof Difficulty[keyof typeof Difficulty];


export const Difficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
  Expert: 'Expert',
} as const;

export type Assessment = typeof Assessment[keyof typeof Assessment];


export const Assessment = {
  SAFE: 'SAFE',
  SUSPICIOUS: 'SUSPICIOUS',
  PHISHING: 'PHISHING',
} as const;

export interface EmailScenario {
  senderName: string;
  senderEmail: string;
  recipientName: string;
  subject: string;
  receivedAt: string;
  body: string;
  /** @nullable */
  displayedUrl: string | null;
  /** @nullable */
  attachment: string | null;
}

export interface Mission {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: Difficulty;
  estimatedMinutes: number;
  xpReward: number;
  scenario: EmailScenario;
  completed: boolean;
  createdAt: string;
}

export type MissionAdmin = Mission & {
  answerAssessment: Assessment;
  correctFindings: string[];
};

export interface MissionInput {
  /**
     * @minLength 3
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 500
     */
  description: string;
  /**
     * @minLength 2
     * @maxLength 60
     */
  category: string;
  difficulty: Difficulty;
  /**
     * @minimum 1
     * @maximum 180
     */
  estimatedMinutes: number;
  /**
     * @minimum 10
     * @maximum 500
     */
  xpReward: number;
  scenario: EmailScenario;
  answerAssessment: Assessment;
  /**
     * @minItems 1
     * @items.minLength 2
     * @items.maxLength 60
     */
  correctFindings: string[];
}

export interface MissionListData {
  missions: Mission[];
}

export interface MissionListResponse {
  success: true;
  data: MissionListData;
}

export interface MissionData {
  mission: Mission;
}

export interface MissionResponse {
  success: true;
  data: MissionData;
}

export interface AdminMissionListData {
  missions: MissionAdmin[];
}

export interface AdminMissionListResponse {
  success: true;
  data: AdminMissionListData;
}

export interface AdminMissionData {
  mission: MissionAdmin;
}

export interface AdminMissionResponse {
  success: true;
  data: AdminMissionData;
}

export interface AttemptInput {
  assessment: Assessment;
  /**
     * @maxItems 8
     * @items.minLength 2
     * @items.maxLength 60
     */
  findings: string[];
}

export type AIAnalysisRiskLevel = typeof AIAnalysisRiskLevel[keyof typeof AIAnalysisRiskLevel];


export const AIAnalysisRiskLevel = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export interface AIAnalysis {
  riskLevel: AIAnalysisRiskLevel;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  correctFindings: string[];
  missedFindings: string[];
  explanation: string;
  recommendations: string[];
  nextRecommendedTopic: string;
  encouragement: string;
}

export interface Attempt {
  id: number;
  missionId: number;
  assessment: Assessment;
  findings: string[];
  score: number;
  createdAt: string;
}

export interface AttemptData {
  attempt: Attempt;
  feedback: AIAnalysis;
  xpAwarded: number;
  completed: boolean;
  nextRecommendedTopic: string;
}

export interface AttemptResponse {
  success: true;
  data: AttemptData;
}

export interface CategoryScore {
  category: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  attempts: number;
}

export interface RecentActivity {
  missionTitle: string;
  category: string;
  score: number;
  xpAwarded: number;
  createdAt: string;
}

export interface Progress {
  xp: number;
  level: number;
  currentLevelXp: number;
  nextLevelXp: number;
  missionsCompleted: number;
  accuracy: number;
  streak: number;
  globalRank: number;
  categoryScores: CategoryScore[];
  recentActivity: RecentActivity[];
  /** @nullable */
  recommendation: string | null;
}

export interface ProgressResponse {
  success: true;
  data: Progress;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  name: string;
  level: number;
  xp: number;
  missions: number;
}

export interface LeaderboardData {
  period: string;
  entries: LeaderboardEntry[];
  /** @nullable */
  currentUserId: number | null;
}

export interface LeaderboardResponse {
  success: true;
  data: LeaderboardData;
}

export interface MentorInput {
  /**
     * @minLength 3
     * @maxLength 1000
     */
  question: string;
}

export interface MentorData {
  answer: string;
  safetyRedirect: boolean;
  topic: string;
}

export interface MentorResponse {
  success: true;
  data: MentorData;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Authentication required
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Administrator access required
 */
export type ForbiddenResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Resource already exists
 */
export type ConflictResponse = ErrorResponse;

export type ListMissionsParams = {
/**
 * @maxLength 100
 */
search?: string;
category?: string;
difficulty?: Difficulty;
sort?: ListMissionsSort;
};

export type ListMissionsSort = typeof ListMissionsSort[keyof typeof ListMissionsSort];


export const ListMissionsSort = {
  recommended: 'recommended',
  title: 'title',
  reward: 'reward',
} as const;

export type GetLeaderboardParams = {
period?: GetLeaderboardPeriod;
};

export type GetLeaderboardPeriod = typeof GetLeaderboardPeriod[keyof typeof GetLeaderboardPeriod];


export const GetLeaderboardPeriod = {
  global: 'global',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;