# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberQuest AI defensive training API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Read the signed-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Read the signed-in user's profile
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateMyProfile
      tags: [users]
      summary: Update the signed-in user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ApiError:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, const: false }
        error: { $ref: "#/components/schemas/ApiError" }
      required: [success, error]
...
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [user, admin] }
        xp: { type: integer }
        level: { type: integer }
        streak: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, name, email, role, xp, level, streak, createdAt]
    UserData:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
      required: [user]
    UserResponse:
      type: object
      properties:
        success: { type: boolean, const: true }
        data: { $ref: "#/components/schemas/UserData" }
      required: [success, data]
    AuthResponse:
      type: object
      properties:
        success: { type: boolean, const: true }
        data: { $ref: "#/components/schemas/UserData" }
      required: [success, data]
    RegisterInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 10, maxLength: 100 }
      required: [name, email, password]
    LoginInput:
      type: object
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 1, maxLength: 100 }
      required: [email, password]
    ProfileUpdateInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
      required: [name]
...