# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberQuest AI defensive training API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiError {
  code: string;
  message: string;
}